Rename file and reset points when resizing
