+++ typescript/web/src/components/labeling-tool/openlayers-map/select-and-modify-feature/resize-box-interaction-iog.ts
